evaluations/requirements.txt
azure-ai-evaluation
azure-identity
